backend/pyproject.toml
[project]
name = "doc-intel-backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "httpx>=0.28.0",
    "python-multipart>=0.0.17",
    "aiofiles>=24.1.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "weaviate-client>=4.9.0",
    "openai>=1.60.0",
    "langgraph-checkpoint-asyncpg",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "langchain-weaviate>=0.0.6",
    "sentence-transformers>=5.3.0",
    "langchain-community>=0.4.1",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.9.0",
]
evals = [
    "langsmith>=0.2.0",
    "ragas>=0.2.0",
    "sqlglot>=26.0.0",
    "arize-phoenix>=6.0.0",
    "numpy>=1.26.0",
]

[tool.uv.sources]
langgraph-checkpoint-asyncpg = { path = "../langgraph-checkpoint-asyncpg", editable = true }

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["N803"]
"src/db/models/**" = ["F821"]
"evals/**" = ["N803"]

[tool.coverage.run]
source = ["src"]

[tool.coverage.report]
fail_under = 80
show_missing = true
